requirements.txt
discord.py==0.16.12
requests==2.11.1
environs==4.0.0
